+++ govtool/frontend/src/components/organisms/DashboardTopNav.tsx
@@ -1,11 +1,15 @@
 import { useEffect, useState } from "react";
-import { Box, Button, IconButton } from "@mui/material";
-import { IconPlusCircle } from "@intersect.mbo/intersectmbo.org-icons-set";
+import { Box, IconButton } from "@mui/material";
 
 import { VotingPowerChips, Typography } from "@atoms";
-import { ICONS, PATHS, PDF_PATHS } from "@consts";
-import { useScreenDimension, useTranslation } from "@hooks";
+import { ICONS } from "@consts";
+import {
+  useGetDRepVotingPowerQuery,
+  useGetVoterInfo,
+  useScreenDimension,
+} from "@hooks";
 import { DashboardDrawerMobile } from "@organisms";
+import { useCardano } from "@context";
 
 type DashboardTopNavProps = {
   title: string;
@@ -18,26 +22,15 @@ export const DashboardTopNav = ({
   title,
   isVotingPowerHidden,
 }: DashboardTopNavProps) => {
-  const { t } = useTranslation();
-  const isProposalDiscussion = Object.values(PDF_PATHS).some(
-    (pdfPath) =>
-      window.location.pathname.includes(pdfPath) &&
-      window.location.pathname.includes(
-        PATHS.connectedProposalPillar.replace("/*", ""),
-      ),
-  );
   const [windowScroll, setWindowScroll] = useState<number>(0);
   const { isMobile } = useScreenDimension();
   const [isDrawerOpen, setIsDrawerOpen] = useState<boolean>(false);
+  const { isEnableLoading } = useCardano();
+  const { voter } = useGetVoterInfo();
+  const { dRepVotingPower } = useGetDRepVotingPowerQuery(voter);
 
   const openDrawer = () => {
     setIsDrawerOpen(true);
-  };
-
-  const goToProposalDiscussionCreateGovernanceAction = () => {
-    window.location.href = `${PATHS.connectedProposalPillar.replace("/*", "")}${
-      PDF_PATHS.proposalDiscussionCreateGovernanceAction
-    }`;
   };
 
   useEffect(() => {
@@ -91,8 +84,16 @@ export const DashboardTopNav = ({
           ) : null}
         </Box>
         <Box display="flex">
-          {!isVotingPowerHidden && !isProposalDiscussion && (
-            <VotingPowerChips />
+          {!isVotingPowerHidden && (
+            <VotingPowerChips
+              isLoading={
+                dRepVotingPower === undefined || !!isEnableLoading || !voter
+              }
+              isShown={
+                voter?.isRegisteredAsDRep || voter?.isRegisteredAsSoleVoter
+              }
+              votingPower={dRepVotingPower}
+            />
           )}
           {isMobile && (
             <IconButton
@@ -102,15 +103,6 @@ export const DashboardTopNav = ({
             >
               <img alt="drawer" src={ICONS.drawerIcon} />
             </IconButton>
-          )}
-          {isProposalDiscussion && (
-            <Button
-              variant="contained"
-              startIcon={<IconPlusCircle fill="white" />}
-              onClick={goToProposalDiscussionCreateGovernanceAction}
-            >
-              {t("proposalDiscussion.proposeAGovernanceAction")}
-            </Button>
           )}
         </Box>
         {isMobile && (